requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
biopython=1.85=py311h9ecbd09_1
cctbx-base=2025.1=pypi_0
gemmi=0.7.0=py311hbebdadd_0
matplotlib-base=3.10.0=py311h2b939e6_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
numpy=1.26.4=py311h24aa872_0
numpy-base=1.26.4=py311hbfb1bba_0
pandas=2.2.3=py311h7db5c69_1
pandoc=3.6.3=ha770c72_0
python=3.11.11=h9e4cc4f_1_cpython
scipy=1.14.0=py311h0a5b728_2
tqdm=4.66.4=py311h92b7b1e_0
